pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "snap-dbx-framework"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "databricks-bundles>=0.275.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["resources*"]
